Improve condition for .yo-rc.json move

+++ generators/app/index.js
@@ -356,7 +356,10 @@ ${dvcRemoteType}://`,
 
   end() {
     // A ".yo-rc.json" file may have been created at the starting path during the prompting step instead of the inside of the generated project.
-    if (this.fs.exists(this.destinationPath(path.join("..", ".yo-rc.json")))) {
+    if (
+      !this.fs.exists(this.destinationPath(".yo-rc.json")) &&
+      this.fs.exists(this.destinationPath(path.join("..", ".yo-rc.json")))
+    ) {
       this.fs.move(
         this.destinationPath(path.join("..", ".yo-rc.json")),
         this.destinationPath(".yo-rc.json")
